Struct factorio_bot_core::factorio::rcon::FactorioRcon

pub struct FactorioRcon { /* private fields */ }

Implementations

impl FactorioRcon

pub async fn new(
settings: &RconSettings,
silent: Arc<RwLock<bool>>
) -> Result<Self>

pub fn new_empty() -> Self

Create a FactorioRcon instance without any connection which would fail if used Why? because LuaRconBuilder requires FactorioRcon which i didnt want to change to an option.

pub async fn initialize_server(&self) -> Result<()>

Executes a couple of commands that need to be send to a newly started factorio server

pub async fn send(&self, command: &str) -> Result<Option<Vec<String>>>

Sends raw command to factorio server

pub async fn screenshot(&self, width: i16, height: i16, depth: i8) -> Result<()>

Take a screenshot -> but where?

pub async fn print(&self, message: &str) -> Result<()>

Print given message to all Clients as Chat Message from Server loudly using /c

pub async fn silent_print(&self, str: &str) -> Result<()>

Print given message to all Clients as Chat Message from Server silenty using /silent-command

pub async fn server_save(&self) -> Result<()>

Save the current game on server

pub async fn whoami(&self, name: &str) -> Result<()>

Starts initial discovery process for “server”

pub async fn add_research(&self, technology_name: &str) -> Result<()>

Adds research to the queue

pub async fn cheat_item(
&self,
player_id: PlayerId,
item_name: &str,
item_count: u32
) -> Result<()>

Cheats in an Item in given quantity to given player

pub async fn cheat_technology(&self, technology_name: &str) -> Result<()>

pub async fn cheat_all_technologies(&self) -> Result<()>

pub async fn place_blueprint(
&self,
player_id: PlayerId,
blueprint: String,
position: &Position,
direction: u8,
force_build: bool,
only_ghosts: bool,
inventory_player_ids: Vec<u8>,
world: &Arc<FactorioWorld>
) -> Result<Vec<FactorioEntity>>

pub async fn revive_ghost(
&self,
player_id: PlayerId,
name: &str,
position: &Position,
world: &Arc<FactorioWorld>
) -> Result<FactorioEntity>

pub async fn cheat_blueprint(
&self,
player_id: PlayerId,
blueprint: String,
position: &Position,
direction: u8,
force_build: bool
) -> Result<Vec<FactorioEntity>>

pub async fn store_map_data(&self, key: &str, value: Value) -> Result<()>

pub async fn retrieve_map_data(&self, key: &str) -> Result<Option<Value>>

pub async fn move_player(
&self,
world: &Arc<FactorioWorld>,
player_id: PlayerId,
goal: &Position,
radius: Option<f64>
) -> Result<()>

pub async fn player_mine(
&self,
world: &Arc<FactorioWorld>,
player_id: PlayerId,
name: &str,
position: &Position,
count: u32
) -> Result<()>

pub async fn player_craft(
&self,
world: &Arc<FactorioWorld>,
player_id: PlayerId,
recipe: &str,
count: u32
) -> Result<()>

pub async fn inventory_contents_at(
&self,
entities: Vec<RequestEntity>
) -> Result<Vec<Option<InventoryResponse>>>

pub async fn player_force(&self) -> Result<FactorioForce>

pub async fn place_entity(
&self,
player_id: PlayerId,
item_name: String,
entity_position: Position,
direction: u8,
world: &Arc<FactorioWorld>
) -> Result<FactorioEntity>

pub async fn insert_to_inventory(
&self,
player_id: PlayerId,
entity_name: String,
entity_position: Position,
inventory_type: u32,
item_name: String,
item_count: u32,
world: &Arc<FactorioWorld>
) -> Result<()>

pub async fn remove_from_inventory(
&self,
player_id: PlayerId,
entity_name: String,
entity_position: Position,
inventory_type: u32,
item_name: String,
item_count: u32,
world: &Arc<FactorioWorld>
) -> Result<()>

pub async fn is_area_empty(&self, area_filter: &AreaFilter) -> Result<bool>

pub async fn find_entities_filtered(
&self,
area_filter: &AreaFilter,
search_name: Option<String>,
search_type: Option<String>
) -> Result<Vec<FactorioEntity>>

pub async fn parse_map_exchange_string(
&self,
name: &str,
map_exchange_string: &str
) -> Result<()>

pub async fn find_tiles_filtered(
&self,
area_filter: &AreaFilter,
name: Option<String>
) -> Result<Vec<FactorioTile>>

pub async fn player_path(
&self,
world: &Arc<FactorioWorld>,
player_id: PlayerId,
goal: &Position,
radius: Option<f64>
) -> Result<Vec<Position>>

pub async fn path(
&self,
world: &Arc<FactorioWorld>,
start: &Position,
goal: &Position,
radius: Option<f64>
) -> Result<Vec<Position>>

pub async fn action_start_walk_waypoints(
&self,
action_id: ActionId,
player_id: PlayerId,
waypoints: Vec<Position>
) -> Result<()>

pub async fn action_start_mining(
&self,
action_id: ActionId,
player_id: PlayerId,
name: &str,
position: &Position,
count: u32
) -> Result<()>

pub async fn plan_path(
&self,
world: &Arc<FactorioWorld>,
entity_name: &str,
entity_type: &str,
underground_entity_name: &str,
underground_entity_type: &str,
underground_max: u8,
from_position: &Position,
to_position: &Position,
to_direction: Direction
) -> Result<Vec<FactorioEntity>>

pub async fn action_start_crafting(
&self,
action_id: ActionId,
player_id: PlayerId,
recipe: &str,
count: u32
) -> Result<()>

pub async fn find_offshore_pump_placement_options(
&self,
world: &Arc<FactorioWorld>,
search_center: Position,
pump_direction: Direction
) -> Result<Vec<Pos>>